DSD-1609: Addressed some code review comments
