+++ src/theme/components/multiSelect.ts
@@ -1,15 +1,16 @@
 export const multiSelectWidths = {
   fitContent: {
-    width: "fit-content",
+    width: 'fit-content',
     minWidth: {
       md: "175px",
     },
     maxWidth: {
       md: "200px",
-    },
+    },      
   },
   full: {
     width: "100%",
+    md: "100%",
   },
 };
 
@@ -20,17 +21,10 @@ const MultiSelect = {
     "menuButton",
     "menuContainer",
     "menuChildren",
-    "viewAllButton",
-    "accordian",
-    "hideAccordian",
+    'viewAllButton',
+    'accordian',
   ],
-  baseStyle: ({
-    isBlockElement,
-    isOpen,
-    width = "fitContent",
-    hasSelectedItems,
-    isDropDownOpen,
-  }) => ({
+  baseStyle: ({ isBlockElement, isOpen, width = "full" }) => ({
     ...multiSelectWidths[width],
     position: "relative",
     actionButtons: {
@@ -42,8 +36,8 @@ const MultiSelect = {
       width: "auto",
     },
     menuButton: {
-      minWidth: "175px",
-      width: "fit-content",
+      minWidth: width === 'full'? '100%': '175px',
+      width: width === 'full'? '100%' : 'fit-content',
     },
     menuContainer: {
       backgroundColor: "white",
@@ -139,28 +133,54 @@ const MultiSelect = {
       marginBottom: 0,
       color: "ui.link.primary",
     },
-    accordianLabel: {
-      marginLeft: hasSelectedItems ? 54 : 0,
+    buttonTextLabel: {
       marginBottom: 0,
+      whiteSpace: 'nowrap',
+      width: '70px',
+      overflow: 'hidden',
+      textOverflow: 'ellipsis',
+      marginLeft: 54
     },
-    menuInputText: {
+    menuSearchInputBox: {
       input: {
         backgroundColor: "ui.bg.default",
         border: "none",
       },
       button: {
-        width: "fit-content",
+        width: width === 'full'? '100%' : 'fit-content',
       },
       minWidth: {
-        md: "175px",
+        md: width === 'full'? '100%': '175px',
       },
       maxWidth: {
         md: "200px",
       },
     },
-    hideAccordian: {
-      display: isDropDownOpen ? "block" : "none",
-    },
+    accordionStyles: {
+      _first: {
+        "> div": {
+          button: {
+            background: "ui.link.primary-05",
+            minWidth: width === 'full' ? '100%' : '175px',
+            width: width === 'full' ? '100%' : 'fit-content',
+          },
+          "> div": {
+            button: { minWidth: "0px" },
+          },
+        },
+      },
+      "> div": {
+        _first: {
+          "> div": {
+            _first: {
+              minWidth: width === 'full' ? "100%" : "200px",
+              maxWidth: width === 'full' ? "100%" : "300px",
+              borderTopWidth: "0.5px",
+            },
+          },
+        },
+      },
+    }
   }),
 };
