+++ src/theme/components/multiSelectMenuButton.ts
@@ -6,7 +6,6 @@ const MultiSelectMenuButton = {
     "menuButton",
     "selectedItemsCountButton",
     "toggleIcon",
-    "menuClearButton",
   ],
   baseStyle: ({ hasSelectedItems = false, isOpen = false }) => ({
     buttonLabel: {
